some more outputs

+++ scaling/length_controlled.py
@@ -113,7 +113,9 @@ def fs_cot(
     max_batch_size,
     stop_strings,
     num_samples,
+    do_sample,
     temperature,
+    num_beams
 ):
     outputs = []
     ex_idx = 0
@@ -131,8 +133,9 @@ def fs_cot(
             tokenizer=tokenizer,
             stop_strings=stop_strings,
             num_return_sequences=num_samples,
-            do_sample=True,
+            do_sample=do_sample,
             temperature=temperature,
+            num_beams=num_beams,
             pad_token_id=tokenizer.eos_token_id,
         )
         model_predictions = tokenizer.batch_decode(
@@ -180,8 +183,8 @@ def load_model(model_name):
     )
     tokenizer.pad_token_id = tokenizer.eos_token_id
     return model, tokenizer
-    
-def run_experiment_temperature_based(model, modelname, domain, max_batch_size, num_samples, temperature, num_ex):
+
+def rum_experiment(model, modelname, domain, max_batch_size, num_samples, temperature, num_beams, num_ex):
     if domain == "trivia_qa":
         examples = load_dataset(
             "mandarjoshi/trivia_qa", "rc", split="validation"
@@ -219,9 +222,10 @@ def run_experiment_temperature_based(model, modelname, domain, max_batch_size, n
             predicted_answer = re.search(r'So the final answer is: (.+)', prediction.strip())
             if predicted_answer:
                 predicted_answer = predicted_answer.group(1).strip().rstrip(".")
-                is_correct = re.sub(
-                    r"[^\w\s]", "", prediction.lower()
-                ) in exs[batch_idx]["answer"]["normalized_aliases"]
+                normalized_answer = re.sub(
+                    r"[^\w\s]", "", predicted_answer.lower()
+                )
+                is_correct = normalized_answer in exs[batch_idx]["answer"]["normalized_aliases"]
             return is_correct, predicted_answer
     if domain in {"indexing", "idx_management", "arrayworld"}:
         from prompts import code_solving_insn, python_exc_cot_query_template, python_exc_cot_insn
@@ -244,15 +248,15 @@ def run_experiment_temperature_based(model, modelname, domain, max_batch_size, n
             f"```\n{code.strip()}\n```" for code in fs_examples
         )
 
-        cot_query_template = ""
-        for ex in indexing_cot_exemplars:
+        cot_query_template = python_exc_cot_insn
+        for ex in cot_exemplars:
             code_assert_prefix = (
                 re.search(assert_regex, ex[0], re.MULTILINE).group(0)
             )
             cot_query_template += "\n\n" + python_exc_cot_query_template.format(code=code_assert_prefix + "??", cot=ex[1])
 
         examples = examples[:min(len(examples), num_ex)]
-        get_batch = lambda examples, ex_idx, max_batch_size: examples[ex_idx : min(len(examples), ex_idx + max_batch_size)]
+        get_batch = lambda ex_idx, max_batch_size: examples[ex_idx : min(len(examples), ex_idx + max_batch_size)]
 
         def make_basic_queries(exs):
             queries = []
@@ -265,7 +269,7 @@ def run_experiment_temperature_based(model, modelname, domain, max_batch_size, n
             return queries
 
         make_cot_queries = lambda exs: [
-            cot_query_template + python_exc_cot_query_template.format(
+            cot_query_template + "\n\n" + python_exc_cot_query_template.format(
                 code=re.search(assert_regex, ex["code"], re.MULTILINE).group(0) + "??",
                 cot=""
             )
@@ -326,7 +330,7 @@ def run_experiment_temperature_based(model, modelname, domain, max_batch_size, n
         examples = json.load(open("data/bamboogle_prerelease.json"))
         examples = examples[:min(len(examples), num_ex)]
 
-        get_batch = lambda examples, ex_idx, max_batch_size: examples[ex_idx : min(len(examples), ex_idx + max_batch_size)]
+        get_batch = lambda ex_idx, max_batch_size: examples[ex_idx : min(len(examples), ex_idx + max_batch_size)]
         make_basic_queries = lambda exs: [
             qa_basic_prompt.format(question=ex["Question"]) for ex in exs
         ]
@@ -371,14 +375,11 @@ def run_experiment_temperature_based(model, modelname, domain, max_batch_size, n
     fs_basic_correct = len([ex for ex in fs_basic_outputs if ex["correct"]])
     print(f"Correct: {fs_basic_correct} / {total_ex}")
     print(
-        f"Avg tot tokens: {sum(op['total_compute_tokens'] for op in fs_basic_outputs) / total_ex:.2f}"
-    )
-    print(
         f"Avg gen tokens: {sum(op['generated_tokens'] for op in fs_basic_outputs) / total_ex:.2f}"
     )
 
-    print(f" ====== FS COT TEMP {temperature}===== ")
-    output_filename = f"outputs/{domain}_fs_cot_temp{temperature*100}_{modelname}.json"
+    print(f" ====== FS COT MULTINOMIAL N={num_samples} TEMP {temperature}===== ")
+    output_filename = f"outputs/{domain}_fs_cot_temp{int(temperature*100)}_N{num_samples}_{modelname}.json"
     if os.path.exists(output_filename):
         fs_cot_outputs = json.load(open(output_filename))
     else:
@@ -392,17 +393,47 @@ def run_experiment_temperature_based(model, modelname, domain, max_batch_size, n
             max_batch_size,
             cot_stop_strings,
             num_samples,
-            temperature
+            True,
+            temperature,
+            1
+        )
+        with open(output_filename, "w") as wf:
+            json.dump(fs_cot_outputs, wf, indent=4)
+
+    fs_cot_correct = len([ex for ex in fs_cot_outputs if any(gen["correct"] for gen in ex['generations'])])
+    print(f"(pass@{num_samples}) Correct: {fs_cot_correct} / {total_ex}")
+    print(
+        f"Avg gen tokens: {sum(gen['generated_tokens'] for op in fs_cot_outputs for gen in op["generations"]) / total_ex:.2f}"
+    )
+
+
+    print(f" ====== FS COT BEAMS {num_beams}===== ")
+    output_filename = f"outputs/{domain}_fs_cot_beams{num_beams}_{modelname}.json"
+    if os.path.exists(output_filename):
+        fs_cot_outputs = json.load(open(output_filename))
+    else:
+        fs_cot_outputs = fs_cot(
+            model,
+            tokenizer,
+            get_batch,
+            make_cot_queries,
+            examples,
+            get_prediction_and_correctness_cot,
+            max_batch_size,
+            cot_stop_strings,
+            1,
+            False,
+            None,
+            num_beams,
         )
         with open(output_filename, "w") as wf:
             json.dump(fs_cot_outputs, wf, indent=4)
     
     fs_cot_correct = len([ex for ex in fs_cot_outputs if any(gen["correct"] for gen in ex['generations'])])
-    print(f"(pass@{num_samples}) Correct: {fs_cot_correct} / {total_ex}")
-    got_correct_lengths = [generation['total_compute_tokens'] for op in fs_cot_outputs for generation in op['generations'] if generation['correct'] ]
-    print("lengths of correct:", got_correct_lengths)
-    got_incorrect_lengths = [generation['total_compute_tokens'] for op in fs_cot_outputs for generation in op['generations'] if not generation['correct'] ]
-    print("lengths of incorrect:", got_incorrect_lengths)
+    print(f"Correct: {fs_cot_correct} / {total_ex}")
+    print(
+        f"Avg gen tokens: {num_beams} * {sum(op["generations"][0]['generated_tokens'] for op in fs_cot_outputs) / total_ex:.2f}"
+    )
 
 
 def run_experiment_exemplar_based(model, modelname, domain, max_batch_size, num_samples, num_ex):
@@ -434,216 +465,24 @@ def run_experiment_exemplar_based(model, modelname, domain, max_batch_size, num_
         basic_stop_strings = ["\nQuestion:"]
         cot_stop_strings = ["\nQuestion:"]
 
-        def get_prediction_and_correctness_basic(prediction, exs, batch_idx):
-            is_correct = re.sub(
-                    r"[^\w\s]", "", prediction.lower()
-                ) in exs[batch_idx]["answer"]["normalized_aliases"]
-            predicted_answer = prediction.lower()
-            return is_correct, predicted_answer
-
-        def get_prediction_and_correctness_cot(prediction, exs, batch_idx):
-            is_correct = False
-            predicted_answer = re.search(r'So the final answer is: (.+)', prediction.strip())
-            if predicted_answer:
-                predicted_answer = predicted_answer.group(1).strip().rstrip(".")
-                is_correct = re.sub(
-                    r"[^\w\s]", "", prediction.lower()
-                ) in exs[batch_idx]["answer"]["normalized_aliases"]
-            return is_correct, predicted_answer
-    if domain in {"indexing", "idx_management", "arrayworld"}:
-        from prompts import code_solving_insn, python_exc_cot_query_template, python_exc_cot_insn
-
-        if domain == "indexing":
-            from prompts import indexing_cot_exemplars as cot_exemplars
-            assert_regex = r"(^[\s\S]*)assert answer == "
-            examples = json.load(open("data/indexing_array_N20.json"))
-        if domain == "idx_management":
-            from prompts import idx_management_cot_exemplars as cot_exemplars
-            assert_regex = r"(^[\s\S]*)assert idx == "
-            examples = json.load(open("data/idx_management_N20.json"))
-        if domain == "arrayworld":
-            from prompts import array_world_cot_exemplars as cot_exemplars
-            assert_regex = r"(^[\s\S]*)assert answer == "
-            examples = json.load(open("data/uniformed_arrayworld_N20.json"))
-
-        fs_examples = [ex[0] for ex in cot_exemplars]
-        code_solving_prompt = code_solving_insn + "\n\n".join(
-            f"```\n{code.strip()}\n```" for code in fs_examples
-        )
-
-        cot_query_template = ""
-        for ex in indexing_cot_exemplars:
-            code_assert_prefix = (
-                re.search(assert_regex, ex[0], re.MULTILINE).group(0)
-            )
-            cot_query_template += "\n\n" + python_exc_cot_query_template.format(code=code_assert_prefix + "??", cot=ex[1])
-
-        examples = examples[:min(len(examples), num_ex)]
-        get_batch = lambda examples, ex_idx, max_batch_size: examples[ex_idx : min(len(examples), ex_idx + max_batch_size)]
-
-        def make_basic_queries(exs):
-            queries = []
-            for ex in exs:
-                code_assert_prefix = (
-                    re.search(assert_regex, ex["code"], re.MULTILINE).group(0).strip()
-                )
-                query = code_solving_prompt + f"\n\n```\n{code_assert_prefix}"
-                queries.append(query)
-            return queries
-
-        make_cot_queries = lambda exs: [
-            cot_query_template + python_exc_cot_query_template.format(
-                code=re.search(assert_regex, ex["code"], re.MULTILINE).group(0) + "??",
-                cot=""
-            )
-            for ex in exs
-        ]
-        basic_stop_strings = ["```"]
-        cot_stop_strings = ["[/ANSWER]"]
-
-        def get_prediction_and_correctness_basic(prediction, exs, batch_idx):
-            assert_line = f"assert answer == {prediction.strip().rstrip('`')}"
-            predicted_answer = None
-            try:
-                tree = ast.parse(assert_line.strip())
-                for node in ast.walk(tree):
-                    if isinstance(node, ast.Assert):
-                        if isinstance(node.test.ops[0], ast.Eq):
-                            answer_node = node.test.comparators[0]
-                            (_, _, predicted_answer) = get_code_str_from_tree_node(answer_node, assert_line)
-                            predicted_answer = predicted_answer.strip()
-                            break
-            except:
-                pass
-            is_correct = False
-            try:
-                is_correct = eval(exs[batch_idx]['true_answer']) == eval(predicted_answer)
-            except: 
-                pass
-
-            return is_correct, predicted_answer
-
-        def get_prediction_and_correctness_cot(prediction, exs, batch_idx):
-            assert_line = re.search(r'\[ANSWER\](.+?)\[/ANSWER\]', prediction)
-            predicted_answer = None
-            if assert_line:
-                assert_line = assert_line.group(1)
-                try:
-                    tree = ast.parse(assert_line.strip())
-                    for node in ast.walk(tree):
-                        if isinstance(node, ast.Assert):
-                            if isinstance(node.test.ops[0], ast.Eq):
-                                answer_node = node.test.comparators[0]
-                                (_, _, predicted_answer) = get_code_str_from_tree_node(answer_node, assert_line)
-                                predicted_answer = predicted_answer.strip()
-                                break
-                except:
-                    pass
-            is_correct = False
-            try:
-                is_correct = (predicted_answer is not None) and (eval(exs[batch_idx]['true_answer']) == eval(predicted_answer))
-            except: 
-                pass
-
-            return is_correct, predicted_answer       
-    if domain == "compgap":
-        from prompts import qa_basic_prompt, compgap_exemplars, qa_cot_template
-        compgap_cot_prompt = "\n\n".join(qa_cot_template.format(question=ex[0], cot=ex[1]) for ex in compgap_exemplars) 
-
-        examples = json.load(open("data/bamboogle_prerelease.json"))
-        examples = examples[:min(len(examples), num_ex)]
-
-        get_batch = lambda examples, ex_idx, max_batch_size: examples[ex_idx : min(len(examples), ex_idx + max_batch_size)]
-        make_basic_queries = lambda exs: [
-            qa_basic_prompt.format(question=ex["Question"]) for ex in exs
-        ]
-        make_cot_queries = lambda exs: [
-            compgap_cot_prompt + qa_cot_template.format(question=ex["Question"], cot="") for ex in exs
-        ]
-        basic_stop_strings = ["\nQuestion:"]
-        cot_stop_strings = ["\nQuestion:"]
-
-        def get_prediction_and_correctness_basic(prediction, exs, batch_idx):
-            is_correct = prediction.lower() == exs[batch_idx]["Answer"].lower()
-            predicted_answer = prediction.lower()
-            return is_correct, predicted_answer
-
-        def get_prediction_and_correctness_cot(prediction, exs, batch_idx):
-            is_correct = False
-            predicted_answer = re.search(r'So the final answer is: (.+)', prediction.strip())
-            if predicted_answer:
-                predicted_answer = predicted_answer.group(1).strip().rstrip(".")
-                is_correct = predicted_answer.lower() == exs[batch_idx]['Answer'].lower()
-            return is_correct, predicted_answer
-
-    total_ex = len(examples)
-
-    print(" ====== FS BASIC GREEDY ===== ")
-    output_filename = f"outputs/{domain}_fs_basic_{modelname}.json"
-    if os.path.exists(output_filename):
-        fs_basic_outputs = json.load(open(output_filename))
-    else:
-        fs_basic_outputs = fs_basic(
-            model,
-            tokenizer,
-            get_batch,
-            make_basic_queries,
-            examples,
-            get_prediction_and_correctness_basic,
-            max_batch_size,
-            basic_stop_strings,
-        )
-        with open(output_filename, "w") as wf:
-            json.dump(fs_basic_outputs, wf, indent=4)
-    fs_basic_correct = len([ex for ex in fs_basic_outputs if ex["correct"]])
-    print(f"Correct: {fs_basic_correct} / {total_ex}")
-    print(
-        f"Avg tot tokens: {sum(op['total_compute_tokens'] for op in fs_basic_outputs) / total_ex:.2f}"
-    )
-    print(
-        f"Avg gen tokens: {sum(op['generated_tokens'] for op in fs_basic_outputs) / total_ex:.2f}"
-    )
-
-    print(f" ====== FS COT TEMP {temperature}===== ")
-    output_filename = f"outputs/{domain}_fs_cot_temp{temperature*100}_{modelname}.json"
-    if os.path.exists(output_filename):
-        fs_cot_outputs = json.load(open(output_filename))
-    else:
-        fs_cot_outputs = fs_cot(
-            model,
-            tokenizer,
-            get_batch,
-            make_cot_queries,
-            examples,
-            get_prediction_and_correctness_cot,
-            max_batch_size,
-            cot_stop_strings,
-            num_samples,
-            temperature
-        )
-        with open(output_filename, "w") as wf:
-            json.dump(fs_cot_outputs, wf, indent=4)
-    
-    fs_cot_correct = len([ex for ex in fs_cot_outputs if any(gen["correct"] for gen in ex['generations'])])
-    print(f"(pass@{num_samples}) Correct: {fs_cot_correct} / {total_ex}")
-    got_correct_lengths = [generation['total_compute_tokens'] for op in fs_cot_outputs for generation in op['generations'] if generation['correct'] ]
-    print("lengths of correct:", got_correct_lengths)
-    got_incorrect_lengths = [generation['total_compute_tokens'] for op in fs_cot_outputs for generation in op['generations'] if not generation['correct'] ]
-    print("lengths of incorrect:", got_incorrect_lengths)
-
-
-
-batch_size = 12
+batch_size = 10
 num_ex = 100
 num_samples = 5
-temperatures = [0.6]
+num_beams = 5
+temperature = 0.6
 
 models = [
     "meta-llama/Llama-3.2-1B-Instruct",
     "meta-llama/Llama-3.2-3B-Instruct",
     "meta-llama/Llama-3.1-8B-Instruct",
 ]
-domains = ["trivia_qa", "compgap", "arrayworld", "indexing", "idx_management"]
+domains = [
+    "trivia_qa", 
+    "compgap", 
+    "arrayworld", 
+    "indexing", 
+    "idx_management"
+]
 
 for model in models:
     modelname = re.search(r"/(.+)", model).group(1)
@@ -651,5 +490,4 @@ for model in models:
     model, tokenizer = load_model(model)
     for domain in domains:
         print(f"---- {domain} ----")
-        for temperature in temperatures:
-            run_experiment_temperature_based(model, modelname, domain, batch_size, num_samples, temperature, num_ex=num_ex)
+        rum_experiment(model, modelname, domain, batch_size, num_samples, temperature, num_beams, num_ex)
